refactor(cli): split main into command modules
- Keep main.py as Typer app mount + version callback only
- Move command groups into cmd_* register modules
- Update architecture tests to scan the CLI package (Refs: #31)
